+++ index.js
@@ -2,9 +2,17 @@ var Expirer = require('expire-unused-keys')
 var xtend = require('xtend')
 var lock = require('level-lock')
 var spaces = require('level-spaces')
+var defaultOpts = {
+	ttl: 3600000,
+	checkInterval: 10000
+}
 
+// https://github.com/rvagg/level-spaces/issues/2
 function isFromThisSpace(key) {
-	return key.indexOf('ÿ') === -1
+	function lacks(char) {
+		return (key.indexOf(char) === -1)
+	}
+	return ( lacks('ÿ') && lacks('~') && lacks('\xff') && lacks('\x00') )
 }
 
 function onCmd(expirer, type, key) {
@@ -17,10 +25,6 @@ function onCmd(expirer, type, key) {
 	}
 }
 
-function onBatch(expirer, cmd) {
-	onCmd(expirer, cmd.type, cmd.key)
-}
-
 function makeChildDb(db) {
 	return db.sublevel ? db.sublevel('expirer') : spaces(db, 'expirer')
 }
@@ -29,14 +33,16 @@ module.exports = function ttl(db, opts) {
 	if (!db) {
 		throw new Error('You must pass a level database to ttl()')
 	}
-	opts = xtend({ttl: 3600000, checkInterval: 10000}, opts)
+	var options = xtend(defaultOpts, opts)
 	var childDb = makeChildDb(db)
-	var expirer = new Expirer(opts.ttl, childDb, opts.checkInterval)
+	var expirer = new Expirer(options.ttl, childDb, options.checkInterval)
 
 	db.on('put', onCmd.bind(null, expirer, 'put'))
 	db.on('del', onCmd.bind(null, expirer, 'del'))
 	db.on('batch', function (cmds) {
-		cmds.forEach(onBatch.bind(null, expirer))
+		cmds.forEach(function (cmd) {
+			onCmd(expirer, cmd.type, cmd.key)
+		})
 	})
 
 	expirer.on('expire', function (key) {
@@ -44,8 +50,8 @@ module.exports = function ttl(db, opts) {
 		if (unlock) {
 			db.del(key, unlock)
 		} else {
-			expirer.touch(key) //retry next time
+			expirer.touch(key) // Retry later
 		}
-	}) //if err, this will probably throw
+	})
 	db.refreshTtl = expirer.touch
 }