+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.0.0
+
+- Added support for `level-spaces@2`
+- The check for whether a key is in a sub level is more thorough.
+
 # 2.0.3
 
 - Updated `.travis.yml`.
@@ -12,7 +17,7 @@
 
 # 2.0.0
 
-- Supporting `level-sublevel`, and `level-spaces`.
+- Added support for `level-spaces@1`.
 
 # 1.1.0
 
@@ -38,4 +43,4 @@
 # 0.0.1
 
 - Initial version.
-- Supporting `level-sublevel`.
+- Added support for `level-sublevel`.